Version 1.1.0

+++ src/globals.hpp
@@ -10,7 +10,7 @@
 namespace globals
 {
 static std::default_random_engine              RANDOM_GENERATOR;
-static std::uniform_int_distribution<uint16_t> X_DISTRIBUTION{ 100, WINDOW_WIDTH - 100 };
-static std::uniform_int_distribution<uint16_t> Y_DISTRIBUTION{ 100, WINDOW_HEIGHT - 100 };
+static std::uniform_int_distribution<uint16_t> X_DISTRIBUTION{ 100, static_cast<uint16_t>(WINDOW_WIDTH - 100) };
+static std::uniform_int_distribution<uint16_t> Y_DISTRIBUTION{ 100, static_cast<uint16_t>(WINDOW_HEIGHT - 100) };
 }  // namespace globals
 #endif  // SRC_GLOBALS_HPP_
